Created Reaping Page.
Created a page for the main game, it also has basic functionality. Right now, it will randomly select an available player and then scan the map around them for other players. It builds an array of nearby players and then uses that data to determine available encounters, and then randomly selects an appropriate encounter.